V 2.0

+++ TermProject/BackEndServer/routes/api/landscapes.js
@@ -3,6 +3,7 @@ let router = express.Router();
 let Landscape = require("../../models/Landscape");
 const multer = require('multer');
 const upload = multer({ storage: multer.memoryStorage() });
+const auth = require("../../middleware/authentication"); // Adjust the path to your auth middleware
 
 router.get("/api/landscapes/:id", async (req, res) => {
     let landscape = await Landscape.findById(req.params.id);
@@ -16,42 +17,44 @@ router.put("/api/landscapes/:id", async (req, res) => {
     await landscape.save();
     return res.send(landscape);
 });
-router.delete("/api/landscapes/:id", async (req, res) => {
-    let landscape = await Landscape.findByIdAndDelete(req.params.id);
-    return res.send(landscape);
+
+router.delete("/api/landscapes/:id", auth, async (req, res) => {
+
+    try {
+        let landscape = await Landscape.findById(req.params.id);
+        if (!landscape) {
+            return res.status(404).send({ message: "Landscape not found" });
+        }
+
+        await Landscape.findByIdAndDelete(req.params.id);
+
+        return res.send(landscape);
+    } catch (error) {
+        return res.status(500).send({ message: "Failed to delete landscape", error: error.message });
+
+    }
 });
 
-// router.post("/api/landscapes", async (req, res) => {
-//     let data = req.body;
-//     let record = new Landscape({
-//         name: data.name,
-//         description: data.description,
-//     });
-//     console.log(req,'request')
-//     try {
-//         await record.save();
-//         console.log(record,req.body)
-//         return res.send(record);
-//     } catch (error) {
-//         console.error('Error saving landscape:', error);
-//         // Respond with a server error status and the error message
-//         return res.status(500).send(error.message);
-//     }
-// });
-
-router.post("/api/landscapes", upload.single('image'), async (req, res) => {
+router.post("/api/landscapes",auth, upload.single('image'), async (req, res) => {
     try {
-        const { name, description } = req.body;
-        const newLandscape = new Landscape({
-            name,
-            description,
-            img: {
-                data: req.file.buffer,
-                contentType: req.file.mimetype
-            }
-        });
-        await newLandscape.save();
-        res.status(201).send(newLandscape);
+
+        if (req.session) {
+            // console.log(req.session , "req.session")
+            const { name, description } = req.body;
+            const newLandscape = new Landscape({
+                name,
+                description,
+                img: {
+                    data: req.file.buffer,
+                    contentType: req.file.mimetype
+                }
+            });
+            await newLandscape.save();
+            res.status(201).send(newLandscape);
+        } else {
+            res.status(401).send({ message: "Unauthorized" });
+        }
+
     } catch (error) {
         res.status(400).send({ message: "Failed to create new landscape", error: error.message });
     }
@@ -69,10 +72,24 @@ router.get('/api/landscapes/image/:id', async (req, res) => {
     }
 });
 
-
 router.get("/api/landscapes", async function (req, res) {
-    let landscapes = await Landscape.find();
-    return res.send(landscapes);
+    try {
+        const page = parseInt(req.query.page) || 1;
+        const limit = parseInt(req.query.limit) || 5;
+        const skip = (page - 1) * limit;
+
+        const landscapes = await Landscape.find().skip(skip).limit(limit);
+        const totalLandscapes = await Landscape.countDocuments();
+
+        return res.json({
+            total: totalLandscapes,
+            page,
+            limit,
+            landscapes
+        });
+    } catch (error) {
+        res.status(500).send("Server error");
+    }
 });
 
 module.exports = router;
